[Add] Add of really basic memory management, and utils folder architecture remake
Memory management is crucial for the code so it must be implemented.
Here is a basic memory management that returns a pointer to the allocated memory.
The allocated memory is defined by an adress (0x10000 arbitrary adress) + memory already used
it's really not perfect and really basic but for now it's working.

Generals:
    - Basic Makefile changes
    - Changes function names in includes files kprint_at_int_at -> kprint_at_int

Utils (src/kernel/utils):
    - Remaking of utils folder architecture, now / only contains .h files that must be only libC api functions
      all the folder in / are lib folder that manage all libs functionnalities
    - for now we only have /string and /memory folder and api