[Add] Branch initialization and add of first shell command
The first shell command is setxkbmap that permit to switch keyboard layout. it's usefull known function that must be added.
for implementation It's necessary to change keyboard drivers for adding a change_lang function that takes a language as parameter.
and change the layout of keyboard in driver. Also for good implementation it was necessary to isolate driver and create a public API
for other programs to use. I think it's a good practice for other drivers to have an external API that only contains usefull variables.

Generals:
    - Makefile basic changes

Utils (src/kernel/utils) :
    - Add of in function for /string library it's checking if char is in string.
    - Add of strcmp function for /string library it's compare two strings and return 1 if they are equal and 0 if they aren't.
    - Add of strcpy function for /string library it's copy a string to another.
    - Add split function for /string library that transforms a string into a list of strings with a delimiter list.
    - Add upper function for /string library that transform a string into upper case.
    - Add of all thoses function into /string api
    - strlen modifications transform parameter string into constant string.

Drivers (src/kernel/drivers) :
    - Creation of keyboard driver API.
    - Add of change_lang function in keyboard driver

Shell (src/kernel/shell) :
    - Creation of builtin_commands folder that must contains all builtin commands.
    - Add of builtin_commands API in shell.
    - Add of setxkbmap function that only takes a language as parameter.